backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0.post1
openai==1.10.0
python-dotenv==1.0.1
requests==2.31.0
llama-index==0.9.39
sqlalchemy==2.0.25
pydantic==2.5.3
google-cloud-firestore==2.14.0
google-generativeai==0.3.2
starlette==0.35.1
pymysql==1.1.0
cryptography==41.0.7